[Metal] Implemented PixelFormatToShaderType. Previously it was a stub. [Metal] Fixed compiler error when using hybrid rendering. [Metal] Fixed compiler error in CopyColourAndDepth_cs.metal

+++ Samples/Media/Hlms/Pbs/Metal/PixelShader_ps.metal
@@ -339,7 +339,7 @@ float4 diffuseCol;
 	@insertpiece( SampleRoughnessMap )
 
 @end @property( hlms_use_prepass )
-	int2 iFragCoord = int2( inPs.gl_FragCoord.xy );
+	ushort2 iFragCoord = ushort2( inPs.gl_FragCoord.xy );
 
 	@property( hlms_use_prepass_msaa )
 		int gBufSubsample = ctz( gl_SampleMask );
